refactor: post 및 s3 디렉토리 domain 디렉토리 안으로 경로 수정
