+++ src/main/java/com/trendist/post_service/domain/post/controller/PostController.java
@@ -1,4 +1,4 @@
-package com.trendist.post_service.post.controller;
+package com.trendist.post_service.domain.post.controller;
 
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -6,9 +6,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.trendist.post_service.global.response.ApiResponse;
-import com.trendist.post_service.post.dto.request.PostCreateRequest;
-import com.trendist.post_service.post.dto.response.PostCreateResponse;
-import com.trendist.post_service.post.service.PostService;
+import com.trendist.post_service.domain.post.dto.request.PostCreateRequest;
+import com.trendist.post_service.domain.post.dto.response.PostCreateResponse;
+import com.trendist.post_service.domain.post.service.PostService;
 
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
