+++ src/main/java/com/trendist/post_service/domain/post/domain/Keyword.java
@@ -1,4 +1,4 @@
-package com.trendist.post_service.post.domain;
+package com.trendist.post_service.domain.post.domain;
 
 public enum Keyword {//게시글 키워드, 협의에 따른 수정 가능
 	Environment,
