+++ src/main/java/com/trendist/post_service/domain/post/repository/PostRepository.java
@@ -1,10 +1,10 @@
-package com.trendist.post_service.post.repository;
+package com.trendist.post_service.domain.post.repository;
 
 import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.trendist.post_service.post.domain.Post;
+import com.trendist.post_service.domain.post.domain.Post;
 
 public interface PostRepository extends JpaRepository<Post, UUID> {
 }
