+++ src/main/java/com/trendist/post_service/domain/s3/controller/PresignedUrlController.java
@@ -1,4 +1,4 @@
-package com.trendist.post_service.s3.controller;
+package com.trendist.post_service.domain.s3.controller;
 
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -6,9 +6,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.trendist.post_service.global.response.ApiResponse;
-import com.trendist.post_service.s3.dto.request.PresignedUrlRequest;
-import com.trendist.post_service.s3.dto.response.PresignedUrlResponse;
-import com.trendist.post_service.s3.service.PresignedUrlService;
+import com.trendist.post_service.domain.s3.dto.request.PresignedUrlRequest;
+import com.trendist.post_service.domain.s3.dto.response.PresignedUrlResponse;
+import com.trendist.post_service.domain.s3.service.PresignedUrlService;
 
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
