+++ src/main/java/com/trendist/post_service/domain/s3/dto/request/PresignedUrlRequest.java
@@ -1,4 +1,4 @@
-package com.trendist.post_service.s3.dto.request;
+package com.trendist.post_service.domain.s3.dto.request;
 
 import java.util.List;
